split write_reg functions into multiple pages
